+++ algorithms/base.py
@@ -30,6 +30,7 @@ GROUP_OPTIMIZE = ("Optimization", "optimization")
 GROUP_EQUITY = ("Equity", "equity")
 GROUP_WALK = ("Walkability", "walkability")
 GROUP_TRANSIT = ("Transit", "transit")
+GROUP_VISIBILITY = ("Visibility", "visibility")
 
 
 class PlanXAlgorithm(QgsProcessingAlgorithm):